Combo
Combined Scripts to decrease "Daisy Chaining" - needed references to the
same scripts, had similar data and functions, etc
Condensed it down to have fewer scripts that are more easily read.
Added new "Ground" prefabs for Level design.
There's probably more useless scripts that can be deleted.
Enemy behaviour is odd, likely to do with the new "Floors"
Mayor "Snaps" to ground being jumped to, this may or may not be a
problem