MASSIVE update - Prototype Phase
Lots of new additions to what we previously had. At least the bare bones
for pretty much everything but bosses.
Next Commit will include a To-Do list.
List of changes in this update:
-Dialogue Scenes have been added
-Dialogue scene activates when the "Token" is touched
-Dialogue scenes currently do not end
-Pause Button Added
-World arranges itself as the Mayor arrives on new platforms
-Mayor now Charges her punches
-uhhhhhh, some other shit too. It's been about a month so I kinda forget
everything that's been done and I've done a poor job keeping track. But
everything stated is kinda the big stuff.
KEY NOTE
This is built with the Unity version 5.4.1
YOU MUST UPDATE UNITY to this version before opening the project. I've
commited the "user preferences" in order to try and auto load Visual
studio prof 2013 instead of the very ugly looking 2015 that doesn't like
to work.