TO-DO List
If you decide to start working on someting from the list:
Inform the team of what item you'll be working on
Create a back up of any prefabs or scripts you'll be working in
Create a new scene (duplicate the LevelTesting scene) to work in.
Once your changes are done Inform the rest of the team so we can
implement
your changes to the main scene and all sync together.